Move types from eth_tx.ts to types.ts
