+++ packages/web3-eth/src/types.ts
@@ -8,135 +8,89 @@ import {
 	HexString32Bytes,
 	HexString256Bytes,
 	HexStringBytes,
-	Uint,
 } from 'web3-utils';
 
-export enum ChainNames {
-	MAINNET = 'mainnet',
-	GOERLI = 'goerli',
-	KOVAN = 'kovan',
-	RINKEBY = 'rinkeby',
-	ROPSTEN = 'ropsten',
+export type chain = 'goerli' | 'kovan' | 'mainnet' | 'rinkeby' | 'ropsten' | 'sepolia';
+export type hardfork =
+	| 'arrowGlacier'
+	| 'berlin'
+	| 'byzantium'
+	| 'chainstart'
+	| 'constantinople'
+	| 'dao'
+	| 'homestead'
+	| 'istanbul'
+	| 'london'
+	| 'merge'
+	| 'muirGlacier'
+	| 'petersburg'
+	| 'shanghai'
+	| 'spuriousDragon'
+	| 'tangerineWhistle';
+
+export interface CustomChain<NumberType = Numbers> {
+	name?: string;
+	networkId: NumberType;
+	chainId: NumberType;
 }
 
-export enum HardForks {
-	CHAIN_START = 'chainstart',
-	HOMESTEAD = 'homestead',
-	DAO = 'dao',
-	TANGERINE_WHISTLE = 'tangerineWhistle',
-	SPURIOUS_DRAGON = 'spuriousDragon',
-	BYZANTIUM = 'byzantium',
-	CONSTANTINOPLE = 'constantinople',
-	PETERSBURG = 'petersburg',
-	ISTANBUL = 'istanbul',
-	BERLIN = 'berlin',
-	LONDON = 'london',
+export interface Common<NumberType = Numbers> {
+	customChain: CustomChain<NumberType>;
+	baseChain?: chain;
+	hardfork?: hardfork;
 }
 
-export interface Transaction {
+export interface Transaction<NumberType = Numbers> {
 	from?: Address;
 	to?: Address;
-	value?: Numbers;
-	gas?: Numbers;
-	gasPrice?: Numbers;
-	type?: Numbers;
-	maxFeePerGas?: Numbers;
-	maxPriorityFeePerGas?: Numbers;
+	value?: NumberType;
+	gas?: NumberType;
+	gasPrice?: NumberType;
+	type?: NumberType;
+	maxFeePerGas?: NumberType;
+	maxPriorityFeePerGas?: NumberType;
 	accessList?: AccessList;
-	input: HexStringBytes;
-	data?: HexString;
-	nonce?: Numbers;
-	chain?: HexString;
-	hardfork?: HexString;
-	common?: {
-		customChain: {
-			name?: string;
-			networkId: Numbers;
-			chainId: Numbers;
-		};
-		baseChain?: ChainNames;
-		hardfork?: HardForks;
-	};
+	data?: HexStringBytes;
+	nonce?: NumberType;
+	chain?: chain;
+	hardfork?: hardfork;
+	chainId?: NumberType;
+	common?: Common<NumberType>;
+	gasLimit?: NumberType;
+	v?: NumberType;
+	r?: HexString;
+	s?: HexString;
 }
 
-interface BaseTransactionFormatted<
-	DesiredType extends ValidTypes = ValidTypes.HexString,
-	ReturnType = ValidReturnTypes[DesiredType],
-> {
-	readonly to?: Address | null;
-	readonly type: ReturnType;
-	readonly nonce: ReturnType;
-	readonly gas: ReturnType;
-	readonly value: ReturnType;
-	readonly input: HexStringBytes;
+export interface PopulatedUnsignedBaseTransaction<NumberType = Numbers> {
+	from: Address;
+	to?: Address;
+	value: Numbers;
+	gas?: Numbers;
+	gasPrice: Numbers;
+	type: Numbers;
+	data: HexStringBytes;
+	nonce: Numbers;
+	chain: chain;
+	hardfork: hardfork;
+	chainId: Numbers;
+	common: Common<NumberType>;
+	gasLimit: Numbers;
 }
-
-interface Transaction1559UnsignedFormatted<
-	DesiredType extends ValidTypes = ValidTypes.HexString,
-	ReturnType = ValidReturnTypes[DesiredType],
-> extends BaseTransactionFormatted<DesiredType> {
-	readonly maxFeePerGas: ReturnType;
-	readonly maxPriorityFeePerGas: ReturnType;
-	readonly accessList: AccessList;
+export interface PopulatedUnsignedEip2930Transaction<NumberType = Numbers>
+	extends PopulatedUnsignedBaseTransaction<NumberType> {
+	accessList: AccessList;
 }
-
-interface Transaction1559SignedFormatted<
-	DesiredType extends ValidTypes = ValidTypes.HexString,
-	ReturnType = ValidReturnTypes[DesiredType],
-> extends Transaction1559UnsignedFormatted<DesiredType> {
-	readonly yParity: ReturnType;
-	readonly r: ReturnType;
-	readonly s: ReturnType;
+export interface PopulatedUnsignedEip1559Transaction<NumberType = Numbers>
+	extends PopulatedUnsignedEip2930Transaction<NumberType> {
+	gasPrice: never;
+	maxFeePerGas: NumberType;
+	maxPriorityFeePerGas: NumberType;
 }
-
-interface Transaction2930UnsignedFormatted<
-	DesiredType extends ValidTypes = ValidTypes.HexString,
-	ReturnType = ValidReturnTypes[DesiredType],
-> extends BaseTransactionFormatted<DesiredType> {
-	readonly gasPrice: ReturnType;
-	readonly accessList: AccessList;
-}
-
-interface Transaction2930SignedFormatted<
-	DesiredType extends ValidTypes = ValidTypes.HexString,
-	ReturnType = ValidReturnTypes[DesiredType],
-> extends Transaction2930UnsignedFormatted<DesiredType> {
-	readonly yParity: ReturnType;
-	readonly r: ReturnType;
-	readonly s: ReturnType;
-}
-
-interface TransactionLegacyUnsignedFormatted<
-	DesiredType extends ValidTypes = ValidTypes.HexString,
-	ReturnType = ValidReturnTypes[DesiredType],
-> extends BaseTransactionFormatted<DesiredType> {
-	readonly gasPrice: ReturnType;
-}
-
-interface TransactionLegacySignedFormatted<
-	DesiredType extends ValidTypes = ValidTypes.HexString,
-	ReturnType = ValidReturnTypes[DesiredType],
-> extends TransactionLegacyUnsignedFormatted<DesiredType> {
-	readonly v: ReturnType;
-	readonly r: Uint;
-	readonly s: Uint;
-}
-
-type TransactionSignedFormatted<DesiredType extends ValidTypes = ValidTypes.HexString> =
-	| Transaction1559SignedFormatted<DesiredType>
-	| Transaction2930SignedFormatted<DesiredType>
-	| TransactionLegacySignedFormatted<DesiredType>;
-
-export type TransactionInfoFormatted<
-	DesiredType extends ValidTypes = ValidTypes.HexString,
-	ReturnType = ValidReturnTypes[DesiredType],
-> = TransactionSignedFormatted<DesiredType> & {
-	readonly blockHash: HexString32Bytes | null;
-	readonly blockNumber: ReturnType | null;
-	readonly from: Address;
-	readonly hash: HexString32Bytes;
-	readonly transactionIndex: ReturnType | null;
-};
+export type PopulatedUnsignedTransaction<NumberType = Numbers> =
+	| PopulatedUnsignedBaseTransaction<NumberType>
+	| PopulatedUnsignedEip2930Transaction<NumberType>
+	| PopulatedUnsignedEip1559Transaction<NumberType>;
 
 export interface ReceiptInfoFormatted<
 	DesiredType extends ValidTypes = ValidTypes.HexString,
