+++ packages/web3-eth/test/unit/populate_transaction.test.ts
@@ -4,12 +4,12 @@ import { BlockTags, ValidTypes } from 'web3-utils';
 import { HttpProvider } from 'web3-providers-http';
 
 import * as rpcMethods from '../../src/rpc_methods';
+import { populateTransaction } from '../../src/eth_tx';
 import {
 	PopulatedUnsignedEip1559Transaction,
 	PopulatedUnsignedEip2930Transaction,
-	populateTransaction,
 	Transaction,
-} from '../../src/eth_tx';
+} from '../../src/types';
 import {
 	Eip1559NotSupportedError,
 	UnableToPopulateNonceError,
